TA-2920 update styles & tests after rebase
