+++ test/jest/client/components/templates/district-office/district-office.test.js
@@ -1,20 +1,24 @@
 import React from 'react'
-import { render, cleanup, waitForElement } from 'react-testing-library'
+import { render, cleanup } from 'react-testing-library'
+import DistrictOffice from 'templates/district-office/district-office.jsx'
 import 'jest-dom/extend-expect'
-
-import DistrictOfficeTemplate from 'templates/district-office/district-office.jsx'
-
-const mockOfficeData = {
-  title: 'State District Office'
-}
 
 afterEach(cleanup)
 
 describe('District Office template', () => {
   it('renders the newsletter sign up component', () => {
-    const { getByTestId } = render(<DistrictOfficeTemplate office={mockOfficeData} />)
+    const mockOfficeData = { title: 'State District Office' }
+    const { getByTestId } = render(<DistrictOffice office={mockOfficeData} />)
 
     const content = getByTestId('office-newsletter')
     expect(content).toBeInTheDocument()
+  })
+
+  it('renders a CTA for any district office', async () => {
+    const mockOfficeData = { title: 'Fearless HQ' }
+    const { getByTestId } = render(<DistrictOffice office={mockOfficeData} />)
+
+    const cta = getByTestId('call-to-action')
+    expect(cta).toBeInTheDocument()
   })
 })
